탱크 : 이동 후 공격이 멈춤 해결,  공격 데미지 수정

+++ Source/Maverick/Public/XR_LSJ/SquadManager.h
@@ -31,9 +31,9 @@ class MAVERICK_API ASquadManager : public AActor, public IIAICommand
 	UPROPERTY()
 	struct FSquadData SquadManagerAbility;
 	//분대 현재 HP
-	float CurrentSquadHp;
+	int32 CurrentSquadHp;
 	//분대 최대 HP
-	float MaxSquadHp; 
+	int32 MaxSquadHp; 
 	//분대의 분대원들
 	TArray<class AAISquad*> SquadArray;
 	//분대원 포지션 목록
